Move Torchlight logic to trait
